+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for the Mapbox Search SDK for Android
 
+## Unreleased
+
+### New features
+- Adding Builder class to `SearchResultMetadata` class
+- Adding `CompletionCallback<Unit>` to the `AnalyticsService` methods
+
 ## 2.5.0-beta.1
 
 ### Mapbox dependencies